+++ docs/changes.rst
@@ -31,6 +31,10 @@ We got an updated :doc:`release roadmap <development/roadmap>`!
     lists of :class:`mopidy.models.Track` instead of
     :class:`mopidy.models.Playlist`, as none of the other fields on the
     ``Playlist`` model was in use.
+  - :meth:`mopidy.backends.BaseCurrentPlaylistController.remove()`` now takes
+    criterias, just like
+    :meth:`mopidy.backends.BaseCurrentPlaylistController.get()``, instead of
+    the track to remove.
 
 
 0.1.0a2 (2010-06-02)